+++ pkg/services/object/get/prm.go
@@ -3,7 +3,6 @@ package getsvc
 import (
 	"context"
 	"crypto/ecdsa"
-	"hash"
 	"io"
 
 	iprotobuf "github.com/nspcc-dev/neofs-node/internal/protobuf"
@@ -26,6 +25,9 @@ type SubmitDataStreamFunc = func(io.ReadCloser)
 type Prm struct {
 	commonPrm
 
+	rng         *object.Range
+	payloadOnly bool
+
 	localGetBuffer         []byte
 	submitLocalGetStreamFn SubmitStreamFunc
 
@@ -44,21 +46,7 @@ type RangePrm struct {
 	forwardRequestFn ForwardRangeRequestFunc
 }
 
-// RangeHashPrm groups parameters of GetRange service call.
-type RangeHashPrm struct {
-	commonPrm
-
-	hashGen func() hash.Hash
-
-	rngs []object.Range
-
-	salt []byte
-
-	forwardedRangeHashResponse [][]byte
-}
-
 type RequestForwarder func(context.Context, coreclient.MultiAddressClient) (*object.Object, error)
-type RangeRequestForwarder func(context.Context, coreclient.MultiAddressClient) ([][]byte, error)
 
 // ForwardHeadRequestFunc sends currently served HEAD request to remote node
 // through passed connection and returns buffered response with requested
@@ -98,8 +86,6 @@ type commonPrm struct {
 
 	raw bool
 
-	rangeForwarder RangeRequestForwarder
-
 	// signerKey is a cached key that should be used for spawned
 	// requests (if any), could be nil if incoming request handling
 	// routine does not include any key fetching operations
@@ -118,9 +104,25 @@ type ObjectWriter interface {
 	ChunkWriter
 }
 
+// HeaderValidator is an optional interface for validating object headers
+// before suppressing them from the response.
+type HeaderValidator interface {
+	ValidateHeader(*object.Object) error
+}
+
 // SetObjectWriter sets target component to write the object.
 func (p *Prm) SetObjectWriter(w ObjectWriter) {
 	p.objWriter = w
+}
+
+// SetRange sets range of the requested payload data.
+func (p *Prm) SetRange(rng *object.Range) {
+	p.rng = rng
+}
+
+// MarkPayloadOnly requests payload without an object header.
+func (p *Prm) MarkPayloadOnly() {
+	p.payloadOnly = true
 }
 
 // SetChunkWriter sets target component to write the object payload range.
@@ -135,28 +137,9 @@ func (p *RangePrm) SetRange(rng *object.Range) {
 	p.rng = rng
 }
 
-// SetRangeList sets a list of object payload ranges.
-func (p *RangeHashPrm) SetRangeList(rngs []object.Range) {
-	p.rngs = rngs
-}
-
-// SetHashGenerator sets constructor of hashing algorithm.
-func (p *RangeHashPrm) SetHashGenerator(v func() hash.Hash) {
-	p.hashGen = v
-}
-
-// SetSalt sets binary salt to XOR object's payload ranges before hash calculation.
-func (p *RangeHashPrm) SetSalt(salt []byte) {
-	p.salt = salt
-}
-
 // SetCommonParameters sets common parameters of the operation.
 func (p *commonPrm) SetCommonParameters(common *util.CommonPrm) {
 	p.common = common
-}
-
-func (p *commonPrm) SetRangeHashRequestForwarder(f RangeRequestForwarder) {
-	p.rangeForwarder = f
 }
 
 // WithAddress sets object address to be read.
@@ -207,6 +190,16 @@ func (p *Prm) WithBuffer(buffer []byte, submitStreamFn SubmitStreamFunc) {
 // GetBuffer returns buffer settings set using [Prm.WithBuffer].
 func (p Prm) GetBuffer() ([]byte, SubmitStreamFunc) {
 	return p.localGetBuffer, p.submitLocalGetStreamFn
+}
+
+// Range returns payload range settings.
+func (p Prm) Range() *object.Range {
+	return p.rng
+}
+
+// PayloadOnly reports whether only payload was requested.
+func (p Prm) PayloadOnly() bool {
+	return p.payloadOnly
 }
 
 // SetRequestForwarder specifies request transport callback to use for receiving